+++ tests/TestCase/Utility/BackupExportTest.php
@@ -24,13 +24,34 @@ namespace MysqlBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\TestSuite\TestCase;
-use MysqlBackup\Test\TestCase\Utility\BackupExport;
+use MysqlBackup\Utility\BackupExport;
+use Reflection\ReflectionTrait;
 
 /**
  * BackupExportTest class
  */
 class BackupExportTest extends TestCase
 {
+    use ReflectionTrait;
+
+    /**
+     * @var \MysqlBackup\Utility\BackupExport
+     */
+    protected $BackupExport;
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->BackupExport = new BackupExport;
+    }
+
     /**
      * Teardown any static object changes and restore them
      * @return void
@@ -38,6 +59,8 @@ class BackupExportTest extends TestCase
     public function tearDown()
     {
         parent::tearDown();
+
+        unset($this->BackupExport);
 
         //Deletes all backups
         foreach (glob(Configure::read('MysqlBackup.target') . DS . '*') as $file) {
@@ -51,17 +74,15 @@ class BackupExportTest extends TestCase
      */
     public function testConstruct()
     {
-        $instance = new BackupExport();
-
-        $connection = $instance->getConnection();
+        $connection = $this->getProperty($this->BackupExport, 'connection');
         $this->assertEquals($connection['scheme'], 'mysql');
         $this->assertEquals($connection['database'], 'test');
         $this->assertEquals($connection['driver'], 'Cake\Database\Driver\Mysql');
 
-        $this->assertNull($instance->getCompression());
-        $this->assertEquals('sql', $instance->getExtension());
-        $this->assertNull($instance->getFilename());
-        $this->assertNull($instance->getRotate());
+        $this->assertNull($this->getProperty($this->BackupExport, 'compression'));
+        $this->assertEquals('sql', $this->getProperty($this->BackupExport, 'extension'));
+        $this->assertNull($this->getProperty($this->BackupExport, 'filename'));
+        $this->assertNull($this->getProperty($this->BackupExport, 'rotate'));
     }
 
     /**
@@ -71,19 +92,17 @@ class BackupExportTest extends TestCase
      */
     public function testCompression()
     {
-        $instance = new BackupExport();
+        $this->BackupExport->compression('bzip2');
+        $this->assertEquals('bzip2', $this->getProperty($this->BackupExport, 'compression'));
+        $this->assertEquals('sql.bz2', $this->getProperty($this->BackupExport, 'extension'));
 
-        $instance->compression('bzip2');
-        $this->assertEquals('bzip2', $instance->getCompression());
-        $this->assertEquals('sql.bz2', $instance->getExtension());
+        $this->BackupExport->compression('gzip');
+        $this->assertEquals('gzip', $this->getProperty($this->BackupExport, 'compression'));
+        $this->assertEquals('sql.gz', $this->getProperty($this->BackupExport, 'extension'));
 
-        $instance->compression('gzip');
-        $this->assertEquals('gzip', $instance->getCompression());
-        $this->assertEquals('sql.gz', $instance->getExtension());
-
-        $instance->compression(false);
-        $this->assertEquals(false, $instance->getCompression());
-        $this->assertEquals('sql', $instance->getExtension());
+        $this->BackupExport->compression(false);
+        $this->assertEquals(false, $this->getProperty($this->BackupExport, 'compression'));
+        $this->assertEquals('sql', $this->getProperty($this->BackupExport, 'extension'));
     }
 
     /**
@@ -94,7 +113,7 @@ class BackupExportTest extends TestCase
      */
     public function testCompressionWithInvalidString()
     {
-        (new BackupExport())->compression('invalidValue');
+        $this->BackupExport->compression('invalidValue');
     }
 
     /**
@@ -105,51 +124,62 @@ class BackupExportTest extends TestCase
      */
     public function testCompressionWithInvalidBool()
     {
-        (new BackupExport())->compression(true);
+        $this->BackupExport->compression(true);
     }
 
     /**
-     * Test for `filename()` method. This also tests for `$compression`
+     * Test for `filename()` method.
+     *
+     * This also tests for patterns and for the `$compression` property.
      * @test
      */
     public function testFilename()
     {
-        $instance = new BackupExport();
+        //`sql` filename
+        $this->BackupExport->filename('backup.sql');
+        $this->assertEquals('/tmp/backups/backup.sql', $this->getProperty($this->BackupExport, 'filename'));
+        $this->assertFalse($this->getProperty($this->BackupExport, 'compression'));
 
-        $instance->filename('backup.sql');
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql', $instance->getFilename());
-        $this->assertFalse($instance->getCompression());
+        //`sql.bz2` filename
+        $this->BackupExport->filename('backup.sql.bz2');
+        $this->assertEquals('/tmp/backups/backup.sql.bz2', $this->getProperty($this->BackupExport, 'filename'));
+        $this->assertEquals('bzip2', $this->getProperty($this->BackupExport, 'compression'));
 
-        $instance->filename('backup.sql.gz');
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql.gz', $instance->getFilename());
-        $this->assertEquals('gzip', $instance->getCompression());
+        //`sql.gz` filename
+        $this->BackupExport->filename('backup.sql.gz');
+        $this->assertEquals('/tmp/backups/backup.sql.gz', $this->getProperty($this->BackupExport, 'filename'));
+        $this->assertEquals('gzip', $this->getProperty($this->BackupExport, 'compression'));
 
-        $instance->filename('backup.sql.bz2');
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql.bz2', $instance->getFilename());
-        $this->assertEquals('bzip2', $instance->getCompression());
+        //Filename with absolute path
+        $this->BackupExport->filename('/tmp/backups/other.sql');
+        $this->assertEquals('/tmp/backups/other.sql', $this->getProperty($this->BackupExport, 'filename'));
+        $this->assertFalse($this->getProperty($this->BackupExport, 'compression'));
 
-        //Absolute path
-        $instance->filename(Configure::read('MysqlBackup.target') . DS . 'other.sql');
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'other.sql', $instance->getFilename());
-        $this->assertFalse($instance->getCompression());
-    }
+        //Compression is ignored, because a filename has been given
+        $this->BackupExport->compression('gzip')->filename('backup.sql.bz2');
+        $this->assertEquals('backup.sql.bz2', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertEquals('bzip2', $this->getProperty($this->BackupExport, 'compression'));
 
-    /**
-     * Test for `filename()` method. This checks that the `filename()` method
-     *  overwrites the `compression()` method
-     * @test
-     */
-    public function testFilenameRewritesCompression()
-    {
-        $instance = new BackupExport();
+        //Compression is ignored, because a filename has been given
+        $this->BackupExport->compression('bzip2')->filename('backup.sql');
+        $this->assertEquals('backup.sql', basename($this->getProperty($this->BackupExport, 'filename')));
+        $this->assertFalse($this->getProperty($this->BackupExport, 'compression'));
 
-        $instance->compression('gzip')->filename('backup.sql.bz2');
-        $this->assertEquals('backup.sql.bz2', basename($instance->getFilename()));
-        $this->assertEquals('bzip2', $instance->getCompression());
+        //Filename with `{$DATABASE}` pattern
+        $this->BackupExport->filename('{$DATABASE}.sql');
+        $this->assertEquals('test.sql', basename($this->getProperty($this->BackupExport, 'filename')));
 
-        $instance->compression('bzip2')->filename('backup.sql');
-        $this->assertEquals('backup.sql', basename($instance->getFilename()));
-        $this->assertFalse($instance->getCompression());
+        //Filename with `{$DATETIME}` pattern
+        $this->BackupExport->filename('{$DATETIME}.sql');
+        $this->assertRegExp('/^[0-9]{14}\.sql$/', basename($this->getProperty($this->BackupExport, 'filename')));
+
+        //Filename with `{$HOSTNAME}` pattern
+        $this->BackupExport->filename('{$HOSTNAME}.sql');
+        $this->assertEquals('localhost.sql', basename($this->getProperty($this->BackupExport, 'filename')));
+
+        //Filename with `{$TIMESTAMP}` pattern
+        $this->BackupExport->filename('{$TIMESTAMP}.sql');
+        $this->assertRegExp('/^[0-9]{10}\.sql$/', basename($this->getProperty($this->BackupExport, 'filename')));
     }
 
     /**
@@ -159,31 +189,10 @@ class BackupExportTest extends TestCase
      */
     public function testFilenameAlreadyExists()
     {
-        (new BackupExport())->filename('backup.sql')->export();
+        $this->BackupExport->filename('backup.sql')->export();
 
         //Again, same filename
-        (new BackupExport())->filename('backup.sql')->export();
-    }
-
-    /**
-     * Test for `filename()` method, with patterns
-     * @test
-     */
-    public function testFilenameWithPatterns()
-    {
-        $instance = new BackupExport();
-
-        $instance->filename('{$DATABASE}.sql');
-        $this->assertEquals('test.sql', basename($instance->getFilename()));
-
-        $instance->filename('{$DATETIME}.sql');
-        $this->assertRegExp('/^[0-9]{14}\.sql$/', basename($instance->getFilename()));
-
-        $instance->filename('{$HOSTNAME}.sql');
-        $this->assertEquals('localhost.sql', basename($instance->getFilename()));
-
-        $instance->filename('{$TIMESTAMP}.sql');
-        $this->assertRegExp('/^[0-9]{10}\.sql$/', basename($instance->getFilename()));
+        $this->BackupExport->filename('backup.sql')->export();
     }
 
     /**
@@ -194,7 +203,7 @@ class BackupExportTest extends TestCase
      */
     public function testFilenameWithInvalidDirectory()
     {
-        (new BackupExport())->filename('noExistingDir' . DS . 'backup.sql');
+        $this->BackupExport->filename('noExistingDir' . DS . 'backup.sql');
     }
 
     /**
@@ -205,7 +214,7 @@ class BackupExportTest extends TestCase
      */
     public function testFilenameWithInvalidExtension()
     {
-        (new BackupExport())->filename('backup.txt');
+        $this->BackupExport->filename('backup.txt');
     }
 
     /**
@@ -214,10 +223,8 @@ class BackupExportTest extends TestCase
      */
     public function testRotate()
     {
-        $instance = new BackupExport();
-
-        $instance->rotate(10);
-        $this->assertEquals(10, $instance->getRotate());
+        $this->BackupExport->rotate(10);
+        $this->assertEquals(10, $this->getProperty($this->BackupExport, 'rotate'));
     }
 
     /**
@@ -228,7 +235,7 @@ class BackupExportTest extends TestCase
      */
     public function testRotateWithInvalidValue()
     {
-        (new BackupExport())->rotate(-1)->export();
+        $this->BackupExport->rotate(-1)->export();
     }
 
     /**
@@ -237,7 +244,7 @@ class BackupExportTest extends TestCase
      */
     public function testStoreAuth()
     {
-        $auth = (new BackupExport())->getAuth();
+        $auth = $this->invokeMethod($this->BackupExport, '_storeAuth');
 
         $this->assertFileExists($auth);
 
@@ -258,9 +265,18 @@ class BackupExportTest extends TestCase
         $bzip2 = Configure::read('MysqlBackup.bin.bzip2');
         $gzip = Configure::read('MysqlBackup.bin.gzip');
 
-        $this->assertEquals($mysqldump . ' --defaults-file=%s %s | ' . $bzip2 . ' > %s', (new BackupExport())->getExecutable('bzip2'));
-        $this->assertEquals($mysqldump . ' --defaults-file=%s %s | ' . $gzip . ' > %s', (new BackupExport())->getExecutable('gzip'));
-        $this->assertEquals($mysqldump . ' --defaults-file=%s %s > %s', (new BackupExport())->getExecutable(false));
+        $this->assertEquals(
+            $mysqldump . ' --defaults-file=%s %s | ' . $bzip2 . ' > %s',
+            $this->invokeMethod($this->BackupExport, '_getExecutable', ['bzip2'])
+        );
+        $this->assertEquals(
+            $mysqldump . ' --defaults-file=%s %s | ' . $gzip . ' > %s',
+            $this->invokeMethod($this->BackupExport, '_getExecutable', ['gzip'])
+        );
+        $this->assertEquals(
+            $mysqldump . ' --defaults-file=%s %s > %s',
+            $this->invokeMethod($this->BackupExport, '_getExecutable', [false])
+        );
     }
 
     /**
@@ -273,7 +289,7 @@ class BackupExportTest extends TestCase
     {
         Configure::write('MysqlBackup.bin.bzip2', false);
 
-        (new BackupExport())->getExecutable('bzip2');
+        $this->invokeMethod($this->BackupExport, '_getExecutable', ['bzip2']);
     }
 
     /**
@@ -286,7 +302,7 @@ class BackupExportTest extends TestCase
     {
         Configure::write('MysqlBackup.bin.gzip', false);
 
-        (new BackupExport())->getExecutable('gzip');
+        $this->invokeMethod($this->BackupExport, '_getExecutable', ['gzip']);
     }
 
     /**
@@ -295,64 +311,41 @@ class BackupExportTest extends TestCase
      */
     public function testExport()
     {
-        $instance = new BackupExport();
-
-        $filename = $instance->compression(false)->export();
+        //Exports with no compression
+        $filename = $this->BackupExport->compression(false)->export();
         $this->assertFileExists($filename);
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql$/', basename($filename));
 
-        $filename = $instance->filename('backup.sql')->export();
-        $this->assertFileExists($filename);
-        $this->assertEquals('backup.sql', basename($filename));
-    }
-
-    /**
-     * Test for `export()` method, with `bzip2` compression
-     * @test
-     */
-    public function testExportBzip2()
-    {
-        $instance = new BackupExport();
-
-        $filename = $instance->compression('bzip2')->export();
+        //Exports with `bzip2` compression
+        $filename = $this->BackupExport->compression('bzip2')->export();
         $this->assertFileExists($filename);
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql\.bz2$/', basename($filename));
 
-        $filename = $instance->filename('backup.sql.bz2')->export();
-        $this->assertFileExists($filename);
-        $this->assertEquals('backup.sql.bz2', basename($filename));
-    }
-
-    /**
-     * Test for `export()` method, with `gzip2` compression
-     * @test
-     */
-    public function testExportGzip()
-    {
-        $instance = new BackupExport();
-
-        $filename = $instance->compression('gzip')->export();
+        //Exports with `gzip` compression
+        $filename = $this->BackupExport->compression('gzip')->export();
         $this->assertFileExists($filename);
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql\.gz$/', basename($filename));
 
-        $filename = $instance->filename('backup.sql.gz')->export();
+        //Exports with `sql` filename
+        $filename = $this->BackupExport->filename('backup.sql')->export();
+        $this->assertFileExists($filename);
+        $this->assertEquals('backup.sql', basename($filename));
+
+        //Exports with `sql.bz2` filename
+        $filename = $this->BackupExport->filename('backup.sql.bz2')->export();
+        $this->assertFileExists($filename);
+        $this->assertEquals('backup.sql.bz2', basename($filename));
+
+        //Exports with `sql.gz` filename
+        $filename = $this->BackupExport->filename('backup.sql.gz')->export();
         $this->assertFileExists($filename);
         $this->assertEquals('backup.sql.gz', basename($filename));
-    }
-
-    /**
-     * Test for `export()` method, with different chmod values
-     * @test
-     */
-    public function testExportWithChmod()
-    {
-        $filename = (new BackupExport())->filename('backup1.sql')->export();
-        $this->assertEquals('0664', substr(sprintf('%o', fileperms($filename)), -4));
 
         //Changes chmod
         Configure::write('MysqlBackup.chmod', 0777);
 
-        $filename = (new BackupExport())->filename('backup2.sql')->export();
+        //Exports with a different chmod
+        $filename = $this->BackupExport->filename('backup2.sql')->export();
         $this->assertEquals('0777', substr(sprintf('%o', fileperms($filename)), -4));
     }
 }