+++ tests/TestCase/Utility/BackupImportTest.php
@@ -24,14 +24,41 @@ namespace MysqlBackup\Test\TestCase\Utility;
 
 use Cake\Core\Configure;
 use Cake\TestSuite\TestCase;
-use MysqlBackup\Test\TestCase\Utility\BackupImport;
 use MysqlBackup\Utility\BackupExport;
+use MysqlBackup\Utility\BackupImport;
+use Reflection\ReflectionTrait;
 
 /**
  * BackupImportTest class
  */
 class BackupImportTest extends TestCase
 {
+    use ReflectionTrait;
+
+    /**
+     * @var \MysqlBackup\Utility\BackupExport
+     */
+    protected $BackupExport;
+
+    /**
+     * @var \MysqlBackup\Utility\$BackupImport
+     */
+    protected $BackupImport;
+
+    /**
+     * Setup the test case, backup the static object values so they can be
+     * restored. Specifically backs up the contents of Configure and paths in
+     *  App if they have not already been backed up
+     * @return void
+     */
+    public function setUp()
+    {
+        parent::setUp();
+
+        $this->BackupExport = new BackupExport;
+        $this->BackupImport = new BackupImport;
+    }
+
     /**
      * Teardown any static object changes and restore them
      * @return void
@@ -39,6 +66,8 @@ class BackupImportTest extends TestCase
     public function tearDown()
     {
         parent::tearDown();
+
+        unset($this->BackupExport, $this->BackupImport);
 
         //Deletes all backups
         foreach (glob(Configure::read('MysqlBackup.target') . DS . '*') as $file) {
@@ -52,9 +81,7 @@ class BackupImportTest extends TestCase
      */
     public function testConstruct()
     {
-        $instance = new BackupImport();
-
-        $connection = $instance->getConnection();
+        $connection = $this->getProperty($this->BackupImport, 'connection');
         $this->assertEquals($connection['scheme'], 'mysql');
         $this->assertEquals($connection['database'], 'test');
         $this->assertEquals($connection['driver'], 'Cake\Database\Driver\Mysql');
@@ -66,33 +93,31 @@ class BackupImportTest extends TestCase
      */
     public function testFilename()
     {
-        $instance = new BackupImport();
-
         //Creates a `sql` backup
-        $backup = (new BackupExport())->filename('backup.sql')->export();
+        $backup = $this->BackupExport->filename('backup.sql')->export();
 
-        $instance->filename($backup);
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql', $instance->getFilename());
-        $this->assertFalse($instance->getCompression());
+        $this->BackupImport->filename($backup);
+        $this->assertEquals('/tmp/backups/backup.sql', $this->getProperty($this->BackupImport, 'filename'));
+        $this->assertFalse($this->getProperty($this->BackupImport, 'compression'));
 
         //Creates a `sql.bz2` backup
-        $backup = (new BackupExport())->filename('backup.sql.bz2')->export();
+        $backup = $this->BackupExport->filename('backup.sql.bz2')->export();
 
-        $instance->filename($backup);
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql.bz2', $instance->getFilename());
-        $this->assertEquals('bzip2', $instance->getCompression());
+        $this->BackupImport->filename($backup);
+        $this->assertEquals('/tmp/backups/backup.sql.bz2', $this->getProperty($this->BackupImport, 'filename'));
+        $this->assertEquals('bzip2', $this->getProperty($this->BackupImport, 'compression'));
 
         //Creates a `sql.gz` backup
-        $backup = (new BackupExport())->filename('backup.sql.gz')->export();
+        $backup = $this->BackupExport->filename('backup.sql.gz')->export();
 
-        $instance->filename($backup);
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql.gz', $instance->getFilename());
-        $this->assertEquals('gzip', $instance->getCompression());
+        $this->BackupImport->filename($backup);
+        $this->assertEquals('/tmp/backups/backup.sql.gz', $this->getProperty($this->BackupImport, 'filename'));
+        $this->assertEquals('gzip', $this->getProperty($this->BackupImport, 'compression'));
 
         //Relative path
-        $instance->filename(basename($backup));
-        $this->assertEquals(Configure::read('MysqlBackup.target') . DS . 'backup.sql.gz', $instance->getFilename());
-        $this->assertEquals('gzip', $instance->getCompression());
+        $this->BackupImport->filename(basename($backup));
+        $this->assertEquals('/tmp/backups/backup.sql.gz', $this->getProperty($this->BackupImport, 'filename'));
+        $this->assertEquals('gzip', $this->getProperty($this->BackupImport, 'compression'));
     }
 
     /**
@@ -103,7 +128,7 @@ class BackupImportTest extends TestCase
      */
     public function testFilenameWithInvalidDirectory()
     {
-        (new BackupImport())->filename('noExistingDir' . DS . 'backup.sql');
+        $this->BackupImport->filename('noExistingDir' . DS . 'backup.sql');
     }
 
     /**
@@ -115,7 +140,8 @@ class BackupImportTest extends TestCase
     public function testFilenameWithInvalidExtension()
     {
         file_put_contents(Configure::read('MysqlBackup.target') . DS . 'backup.txt', null);
-        (new BackupImport())->filename('backup.txt');
+
+        $this->BackupImport->filename('backup.txt');
     }
 
     /**
@@ -124,7 +150,7 @@ class BackupImportTest extends TestCase
      */
     public function testStoreAuth()
     {
-        $auth = (new BackupImport())->getAuth();
+        $auth = $this->invokeMethod($this->BackupImport, '_storeAuth');
 
         $this->assertFileExists($auth);
 
@@ -145,9 +171,18 @@ class BackupImportTest extends TestCase
         $bzip2 = Configure::read('MysqlBackup.bin.bzip2');
         $gzip = Configure::read('MysqlBackup.bin.gzip');
 
-        $this->assertEquals($bzip2 . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s', (new BackupImport())->getExecutable('bzip2'));
-        $this->assertEquals($gzip . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s', (new BackupImport())->getExecutable('gzip'));
-        $this->assertEquals('cat %s | ' . $mysql . ' --defaults-extra-file=%s %s', (new BackupImport())->getExecutable(false));
+        $this->assertEquals(
+            $bzip2 . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s',
+            $this->invokeMethod($this->BackupImport, '_getExecutable', ['bzip2'])
+        );
+        $this->assertEquals(
+            $gzip . ' -dc %s | ' . $mysql . ' --defaults-extra-file=%s %s',
+            $this->invokeMethod($this->BackupImport, '_getExecutable', ['gzip'])
+        );
+        $this->assertEquals(
+            'cat %s | ' . $mysql . ' --defaults-extra-file=%s %s',
+            $this->invokeMethod($this->BackupImport, '_getExecutable', [false])
+        );
     }
 
     /**
@@ -160,7 +195,7 @@ class BackupImportTest extends TestCase
     {
         Configure::write('MysqlBackup.bin.bzip2', false);
 
-        (new BackupImport())->getExecutable('bzip2');
+        $this->invokeMethod($this->BackupImport, '_getExecutable', ['bzip2']);
     }
 
     /**
@@ -173,7 +208,7 @@ class BackupImportTest extends TestCase
     {
         Configure::write('MysqlBackup.bin.gzip', false);
 
-        (new BackupImport())->getExecutable('gzip');
+        $this->invokeMethod($this->BackupImport, '_getExecutable', ['gzip']);
     }
 
     /**
@@ -182,35 +217,21 @@ class BackupImportTest extends TestCase
      */
     public function testImport()
     {
-        //Exports and imports a `sql` backup
-        $backup = (new BackupExport())->compression(false)->export();
-        $filename = (new BackupImport())->filename($backup)->import();
+        //Exports and imports with no compression
+        $backup = $this->BackupExport->compression(false)->export();
+        $filename = $this->BackupImport->filename($backup)->import();
 
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql$/', basename($filename));
-    }
 
-    /**
-     * Test for `import()` method, with `bzip2` compression
-     * @test
-     */
-    public function testImportBzip2()
-    {
-        //Exports and imports a `sql` backup
-        $backup = (new BackupExport())->compression('bzip2')->export();
-        $filename = (new BackupImport())->filename($backup)->import();
+        //Exports and imports with `bzip2` compression
+        $backup = $this->BackupExport->compression('bzip2')->export();
+        $filename = $this->BackupImport->filename($backup)->import();
 
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql\.bz2$/', basename($filename));
-    }
 
-    /**
-     * Test for `import()` method, with `gzip` compression
-     * @test
-     */
-    public function testImportGzip()
-    {
-        //Exports and imports a `sql` backup
-        $backup = (new BackupExport())->compression('gzip')->export();
-        $filename = (new BackupImport())->filename($backup)->import();
+        //Exports and imports with `gzip` compression
+        $backup = $this->BackupExport->compression('gzip')->export();
+        $filename = $this->BackupImport->filename($backup)->import();
 
         $this->assertRegExp('/^backup_test_[0-9]{14}\.sql\.gz$/', basename($filename));
     }
@@ -223,6 +244,6 @@ class BackupImportTest extends TestCase
      */
     public function testImportWithoutFilename()
     {
-        (new BackupImport())->import();
+        $this->BackupImport->import();
     }
 }