+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 # 1.x branch
 ## 1.0 branch
+### 1.0.1
+* added `BackupManager::deleteAll()` and `BackupShell::deleteAll()` methods;
+* improved tests. Also errors in the shell are checked.
+
 ### 1.0.0
 * first release.